Internals

RobustOptLagrangianDual.AbstractProblem — Type

AbstractProblem

Abstract supertype for the definition of specific problems.

RobustOptLagrangianDual.MasterType — Type

MasterType

Enum representing the master problem type. Possible values are CCG and Benders.

RobustOptLagrangianDual.SubproblemType — Type

SubproblemType

Enum representing the subproblem type. Possible values are LinearizedKKT, IndicatorKKT, LinearizedDual, IndicatorDual and LagrangianDual.

RobustOptLagrangianDual.build_checking_sp — Function

build_checking_sp(problem::AbstractProblem, MP::JuMP.Model)

Return the JuMP model of the optimality type subproblem, which also computes an optimal Lagrangian coefficient based on the optimal values of the solved master problem MP.

Must be implemented if mixed_integer_recourse(problem) == false.

RobustOptLagrangianDual.build_feasibility_sp — Function

build_feasibility_sp(problem::AbstractProblem, MP::JuMP.Model, subproblem::SubproblemType)

Return the JuMP model of the feasibility type subproblem based on the optimal value of the master model MP.

Must be implemented if mixed_integer_recourse(problem) == false && complete_recourse(problem) == false.

RobustOptLagrangianDual.build_master_inner_level_check — Function

build_master_inner_level_check(problem::AbstractProblem, MP_outer::JuMP.Model, discrete_decision_list::Dict)

Return the JuMP model of the inner-level master problem, which also computes an optimal Lagrangian coefficient, based on the optimal values of the solved outer-level master problem MP_outer.

Must be implemented if mixed_integer_recourse(problem) = true.

RobustOptLagrangianDual.build_second_stage_problem — Function

build_second_stage_problem(problem::AbstractProblem, MP_outer::JuMP.Model, MP_inner::JuMP.Model)

Return the JuMP model of the deterministic second-stage problem based on the optimal values of the outer-level MP_outer and inner-level MP_inner master models.

Must be implemented if mixed_integer_recourse(problem) == true.

RobustOptLagrangianDual.build_sp — Function

build_sp(problem::AbstractProblem, MP::JuMP.Model, subproblem::SubproblemType, λ::Float64 = 1.0)

Return the JuMP model of the optimality type subproblem based on the optimal value of the master model MP and the Lagrangian coefficient λ.

Must be implemented if mixed_integer_recourse(problem) == false.

RobustOptLagrangianDual.complete_recourse — Function

complete_recourse(problem::AbstractProblem)

Return true if problem satisfies complete recourse and false otherwise.

RobustOptLagrangianDual.compute_lagrangian_coefficient — Function

function compute_lagrangian_parameter(problem::AbstractProblem, MP::JuMP.Model)

Return an optimal Lagrangian coefficient based on the optimal value of the (outer-level) master model MP.

Must be implemented if indicator_uncertainty(problem) == false.

RobustOptLagrangianDual.indicator_uncertainty — Function

indicator_uncertainty(problem::AbstractProblem)

Return true if problem has indicator uncertainties only and false otherwise.

RobustOptLagrangianDual.init_lagrangian_coefficient — Function

function init_lagrangian_coefficient(problem::AbstractProblem, P::JuMP.Model)

Return an optimal Lagrangian coefficient computing using the solved JuMP model P built using either build_checking_sp if mixed_integer_recourse(problem) == false or build_master_inner_level_check if mixed_integer_recourse(problem) == true.

RobustOptLagrangianDual.init_master — Function

init_master(problem::AbstractProblem)

Initialize the (outer-level) master model with variables and constraints.

RobustOptLagrangianDual.init_master_inner_level — Function

init_master_inner_level(problem::AbstractProblem)
init_master_inner_level(problem::AbstractProblem, MP_inner::JuMP.Model)
init_master_inner_level(problem::AbstractProblem, MP_outer::JuMP.Model, discrete_decision_list::Dict, master_inner::SubproblemType, λ = nothing)

Initialize the inner-level master model with variables and constraints.

Must be implemented if mixed_integer_recourse(problem) == true.

In the second version, the variable values must be fixed to optimal values of the solved inner-level master model MP_inner.

In the third version, the model must be built using the first-stage decisions in the solved outer-level master model MP_outer, the discrete second-stage decisions stored in the discrete_decision_list dictionary, and the master_inner type of the inner-level master model. The Lagrangian coefficient λ must be Float64 if master_inner == LagrangianDual.

RobustOptLagrangianDual.mixed_integer_recourse — Function

mixed_integer_recourse(problem::AbstractProblem)

Return true if problem has mixed-integer recourse variables and false otherwise.

RobustOptLagrangianDual.objective_scale — Function

objective_scale(problem::AbstractProblem)

Return the objective scaling factor of problem.

RobustOptLagrangianDual.record_discrete_second_stage_decision — Function

record_discrete_second_stage_decision(problem::AbstractProblem, SP::JuMP.Model, discrete_decision_list::Dict)

Record the optimal discrete second-stage decisions from the solved subproblem model SP in the discrete_decision_list dictionary.

Return true if the decision already existed in discrete_decision_list and false otherwise.

Must be implemented if mixed_integer_recourse(problem) == true.

RobustOptLagrangianDual.record_scenario — Function

record_scenario(problem::AbstractProblem, SP::JuMP.Model, scenario_list::Dict)

This is only for debugging purposes. Record the optimal value of the subproblem model SP in the scenario_list dictionary.

Return true if the scenario already existed in scenario_list and false otherwise.

RobustOptLagrangianDual.solve_deterministic_problem — Function

solve_deterministic_problem(problem::AbstractProblem)

Solve deterministic problem for nominal parameter values.

Return the optimal objective value.

RobustOptLagrangianDual.solve_second_stage_problem_lagrangian — Function

solve_second_stage_problem_lagrangian(problem::AbstractProblem, MP::JuMP.Model, SP::JuMP.Model, λ::Float64)

Solve the second-stage Lagrangian problem with coefficient λ based on the optimal values of the master model MP and subproblem SP.

Return the optimal value of the Lagrangian penalty term.

RobustOptLagrangianDual.update_master_continous — Function

update_master_continuous(problem::AbstractProblem, MP::JuMP.Model, SP::JuMP.Model, master::MasterType, subproblem::SubproblemType)

Update the master model MP based on the optimal value of the continuous subproblem model SP, the master type and subproblem type.

Must be implemented if mixed_integer_recourse(problem) == false.

RobustOptLagrangianDual.update_master_inner_level — Function

update_master_inner_level(problem::AbstractProblem, MP_inner::JuMP.Model, MP_outer::JuMP.Model, SP::JuMP.Model, master_inner::SubproblemType, λ = nothing)

Update the inner-level master model MP_inner based on the optimal values of the outer-level master model MP_outer, the subproblem model SP, and the master_inner type of the inner-level master model.

The Lagrangian coefficient λ must be Float64 if master_inner == LagrangianDual.

Must be implemented if mixed_integer_recourse(problem) = true.

RobustOptLagrangianDual.update_master_mixed_integer — Function

update_master_mixed_integer(problem::AbstractProblem, MP_outer::JuMP.Model, MP_inner::JuMP.Model, master::MasterType)

Update the outer-level master model MP_outer based on the optimal value of the mixed-integer inner-level master model MP_inner and the master type of the inner-level master model.

Must be implemented if mixed_integer_recourse(problem) = true.