Solvers

RobustOptLagrangianDual.set_num_threads — Method

set_num_threads(threadlim::Int)

Set the number of threads to use for parallel branch & bound.

RobustOptLagrangianDual.set_solver_CPLEX — Method

set_solver_CPLEX()

Set the CPLEX solver. CPLEX must be separately installed and added via pkg> add CPLEX. The package defaults to Cbc if this function throws an error.

RobustOptLagrangianDual.set_solver_Cbc — Method

set_solver_Cbc()

Set the Cbc solver.

RobustOptLagrangianDual.set_solver_Gurobi — Method

set_solver_Gurobi()

Set the Gurobi solver. Gurobi must be separately installed and added via pkg> add Gurobi. The package defaults to Cbc if this function throws an error.

RobustOptLagrangianDual.set_solver_Mosek — Method

set_solver_Mosek()

Set the Mosek solver. Mosek must be separately installed and added via pkg> add Mosek, MosekTools. The package defaults to Cbc if this function throws an error.